+++ lib/src/cursor.dart
@@ -1,7 +1,9 @@
 /* This is free and unencumbered software released into the public domain. */
 
-import 'package:flutter_android/android_database.dart' show MatrixCursor;
+import 'package:flutter_android/android_database.dart' show Cursor, MatrixCursor;
 import 'package:meta/meta.dart' show experimental, required;
+
+import 'exceptions.dart';
 
 /// Exposes results from a query on a [SQLiteDatabase].
 ///
@@ -19,4 +21,20 @@ class SQLiteCursor extends MatrixCursor {
   @experimental
   SQLiteCursor.from({@required List<String> columns, @required List<List<dynamic>> rows})
     : super.from(columns: columns, rows: rows);
+
+  /// Returns the value of the requested column as a UTC instant in time.
+  ///
+  /// See: https://www.sqlite.org/lang_datefunc.html
+  DateTime getDateTime(final int columnIndex) {
+    switch (getType(columnIndex)) {
+      case Cursor.FIELD_TYPE_NULL:
+        return null;
+      case Cursor.FIELD_TYPE_INTEGER:
+        return DateTime.fromMillisecondsSinceEpoch(getInt(columnIndex), isUtc: true);
+      case Cursor.FIELD_TYPE_STRING:
+        return DateTime.parse(getString(columnIndex)); // TODO: default timezone?
+      default:
+        throw SQLiteDatatypeMismatchException();
+    }
+  }
 }